style: copy getCreateTableStmt
