+++ sqle/driver/mysql/context/context.go
@@ -418,3 +418,38 @@ func (c *Context) GetSystemVariable(name string) (string, error) {
 	c.AddSysVar(name, value.String)
 	return value.String, nil
 }
+
+// getCreateTableStmt get create table stmtNode for db by query; if table not exist, return null.
+func (i *Inspect) getCreateTableStmt(stmt *ast.TableName) (*ast.CreateTableStmt, bool, error) {
+	exist, err := i.Ctx.IsTableExist(stmt)
+	if err != nil {
+		return nil, exist, err
+	}
+	if !exist {
+		return nil, exist, nil
+	}
+
+	info, _ := i.Ctx.GetTableInfo(stmt)
+	if info.MergedTable != nil {
+		return info.MergedTable, exist, nil
+	}
+	if info.OriginalTable != nil {
+		return info.OriginalTable, exist, nil
+	}
+
+	// create from connection
+	conn, err := i.getDbConn()
+	if err != nil {
+		return nil, exist, err
+	}
+	createTableSql, err := conn.ShowCreateTable(util.GetTableNameWithQuote(stmt))
+	if err != nil {
+		return nil, exist, err
+	}
+	createStmt, err := util.ParseCreateTableStmt(createTableSql)
+	if err != nil {
+		return nil, exist, err
+	}
+	info.OriginalTable = createStmt
+	return createStmt, exist, nil
+}
